+++ src/Renderers/HelloRenderer.h
@@ -40,10 +40,10 @@ class HelloRenderer : public IRenderer {
 
     struct Drawable {
         Buffer VertexBuffer;
-        size_t VertexCount;
+        uint32_t VertexCount;
 
         Buffer IndexBuffer;
-        size_t IndexCount;
+        uint32_t IndexCount;
     };
 
     struct Mesh {
